solar_system_simulator/requirements.txt
pygame
PyOpenGL
numpy
requests
